+++ api/src/resources/attachments.rs
@@ -1,3 +1,4 @@
+use crate::AttachmentReference;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
@@ -6,4 +7,15 @@ pub struct ContentHash(pub String);
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct UploadAttachmentResponse {
     pub content_hash: ContentHash,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, Eq)]
+pub struct AttachmentMetadata {
+    pub name: String,
+    pub size: u64,
+    pub content_type: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub attachment_reference: Option<AttachmentReference>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub content_hash: Option<ContentHash>,
 }