+++ cli/src/commands/create/user.rs
@@ -1,6 +1,6 @@
 use anyhow::{Context, Result};
 use reinfer_client::{
-    Client, Email, GlobalPermission, NewUser, ProjectName, ProjectPermission, Username,
+    Client, GlobalPermission, NewUser, ProjectName, ProjectPermission, UserEmail, Username,
 };
 use std::collections::hash_map::HashMap;
 use structopt::StructOpt;
@@ -15,7 +15,7 @@ pub struct CreateUserArgs {
 
     #[structopt(name = "email")]
     /// Email address of the new user
-    email: Email,
+    email: UserEmail,
 
     #[structopt(long = "global-permissions")]
     /// Global permissions to give to the new user